cleanup of print out commands
